finished ch 7

+++ DogGo/Controllers/DogsController.cs
@@ -3,6 +3,8 @@ using System;
 using System.Collections.Generic;
 using DogGo.Repositories;
 using DogGo.Models;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
 
 namespace DogGo.Controllers
 {
@@ -17,9 +19,11 @@ namespace DogGo.Controllers
         }
 
         // GET: DogsController
+        [Authorize]
         public ActionResult Index()
         {
-            List<Dog> dogs = _dogRepo.GetAllDogs();
+            int ownerId = GetCurrentUserId();
+            List<Dog> dogs = _dogRepo.GetDogsByOwnerId(ownerId);
 
             return View(dogs);
         }
@@ -33,6 +37,7 @@ namespace DogGo.Controllers
         }
 
         // GET: DogsController/Create
+        [Authorize]
         public ActionResult Create()
         {
             return View();
@@ -44,7 +49,9 @@ namespace DogGo.Controllers
         public ActionResult Create(Dog dog)
         {
             try
-            { 
+            {
+                // update the dogs OwnerId to the current user's Id 
+                dog.OwnerId = GetCurrentUserId();
                 _dogRepo.AddDog(dog);
                 return RedirectToAction("Index");
             }
@@ -58,7 +65,10 @@ namespace DogGo.Controllers
         public ActionResult Edit(int id)
         {
             Dog dog = _dogRepo.GetDogById(id);
-            if (dog == null)
+            if (dog.OwnerId != GetCurrentUserId())
+            {
+                return NotFound();
+            } else if (dog == null)
             {
                 return NotFound();
             }
@@ -70,6 +80,10 @@ namespace DogGo.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit(int id, Dog dog)
         {
+            if (dog.OwnerId != GetCurrentUserId())
+            {
+                return NotFound();
+            }
             try
             {
                 _dogRepo.UpdateDog(dog);
@@ -85,6 +99,10 @@ namespace DogGo.Controllers
         public ActionResult Delete(int id)
         {
             Dog dog = _dogRepo.GetDogById(id);
+            if (dog.OwnerId != GetCurrentUserId())
+            {
+                return NotFound();
+            }
             return View(dog);
         }
 
@@ -93,6 +111,10 @@ namespace DogGo.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Delete(int id, Dog dog)
         {
+            if (dog.OwnerId != GetCurrentUserId())
+            {
+                return NotFound();
+            }
             try
             {
                 _dogRepo.DeleteDog(id);
@@ -102,6 +124,12 @@ namespace DogGo.Controllers
             {
                 return View(dog);
             }
+        }
+
+        private int GetCurrentUserId()
+        {
+            string id = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            return int.Parse(id);
         }
     }
 }
